+++ addons/aiCfgFixes/CfgAmmo.hpp
@@ -469,14 +469,5 @@ class CfgAmmo {
         allowAgainstInfantry = 1;
         cost = 50;
     };
-    class potato_R_PG7VL_AA: CUP_R_PG7VL_AT {
-        airLock = 1;
-    };
-    class potato_R_PG7VM_AA: CUP_R_PG7VM_AT {
-        airLock = 1;
-    };
-    class potato_R_PG7V_AA: CUP_R_PG7V_AT {
-        airLock = 1;
-    };
 };
