+++ addons/customGear/launchers/CfgMagazineWells.hpp
@@ -1,0 +1,5 @@
+class CfgMagazineWells {
+    class RPG7 {
+        potato_airlock[] = {"potato_PG7V_M","potato_PG7VM_M","potato_PG7VL_M"};
+    };
+};
